+++ dart/geobase/lib/src/meta/crs/coord_ref_sys.dart
@@ -1,0 +1,139 @@
+// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Use of this source code is governed by a “BSD-3-Clause”-style license that is
+// specified in the LICENSE file.
+//
+// Docs: https://github.com/navibyte/geospatial
+
+// ignore_for_file: constant_identifier_names
+
+import 'package:meta/meta.dart';
+
+import '/src/codes/axis_order.dart';
+
+import 'coord_ref_sys_resolver.dart';
+
+/// Metadata about a coordinate reference system (CRS) identified and specified
+/// by [id].
+///
+/// NOTE: The current version of this class does not yet provide very much CRS
+/// metadata, but in future this might also provide more information.
+@immutable
+class CoordRefSys {
+  /// Metadata about a coordinate reference system (CRS) identified and
+  /// specified by [id].
+  ///
+  /// No normalization of identifiers is done, so for example
+  /// `CoordRefSys.id('http://www.opengis.net/def/crs/EPSG/0/4326')` and
+  /// `CoordRefSys.id('EPSG:4326')` would be different instances even if
+  /// referring to the same coordinate reference system (WGS84
+  /// latitude/longitude).
+  ///
+  /// See also [CoordRefSys.normalized].
+  const CoordRefSys.id(this.id);
+
+  /// Metadata about a coordinate reference system (CRS) identified and
+  /// specified by the normalized identifier of [id].
+  ///
+  /// Normalization: `CoordRefSysResolver.registry.normalizeId(id)`.
+  /// 
+  /// Using the basic default implementation of CoordRefSysResolver
+  /// `CoordRefSys.normalized('http://www.opengis.net/def/crs/EPSG/0/4326')` and
+  /// `CoordRefSys.normalized('EPSG:4326')` would refer to the same instance
+  /// with id normalized as `http://www.opengis.net/def/crs/EPSG/0/4326`.
+  CoordRefSys.normalized(String id)
+      : id = CoordRefSysResolver.registry.normalizeId(id);
+
+  /// The coordinate reference system (CRS) identifier.
+  ///
+  /// The identifier is authorative, it identifies a well known or referenced
+  /// specification that defines properties for a coordinate reference system.
+  ///
+  /// Examples:
+  /// * `http://www.opengis.net/def/crs/OGC/1.3/CRS84`: WGS 84 geographic
+  ///    coordinates (longitude, latitude) ordered as specified by
+  ///    `AxisOrder.xy`.
+  /// * `http://www.opengis.net/def/crs/OGC/1.3/CRS84h`: WGS 84 geographic
+  ///    coordinates (longitude, latitude) ordered as specified by
+  ///    `AxisOrder.xy`, with ellipsoidal height (elevation) as a third
+  ///    coordinate.
+  /// * `http://www.opengis.net/def/crs/EPSG/0/4326` or (`EPSG:4326`): WGS 84
+  ///    geographic coordinates (latitude, longitude) ordered as specified by
+  ///    `AxisOrder.yx`.
+  /// * `http://www.opengis.net/def/crs/EPSG/0/3857` or (`EPSG:3857`): WGS 84
+  ///    projected (Web Mercator) metric coordinates ordered as specified by
+  ///    `AxisOrder.xy`.
+  /// * `http://www.opengis.net/def/crs/EPSG/0/3395` or (`EPSG:3395`): WGS 84
+  ///    projected (World Mercator) metric coordinates ordered as specified by
+  ///    `AxisOrder.xy`.
+  final String id;
+
+  /// Try to resolve an axis order of coordinate values in position and point
+  /// representations for this coordinate reference system identified and
+  /// specified by [id].
+  ///
+  /// The `null` return value is interpreted as "the axis order is not known".
+  AxisOrder? get axisOrder => CoordRefSysResolver.registry.axisOrder(id);
+
+  /// The coordinate reference system resolved in this order:
+  /// 1. [coordRefSys] if it's non-null
+  /// 2. otherwise `CoordRefSys.normalized(crs)` if [crs] is non-null
+  /// 3. otherwise `CoordRefSys.CRS84`
+  factory CoordRefSys.from({
+    CoordRefSys? coordRefSys,
+    String? crs,
+  }) =>
+      coordRefSys ?? (crs != null ? CoordRefSys.normalized(crs) : CRS84);
+
+  /// The coordinate reference system identified by
+  /// 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'.
+  ///
+  /// References WGS 84 geographic coordinates (longitude, latitude) ordered as
+  /// specified by `AxisOrder.xy`.
+  static const CoordRefSys CRS84 =
+      CoordRefSys.id('http://www.opengis.net/def/crs/OGC/1.3/CRS84');
+
+  /// The coordinate reference system identified by
+  /// 'http://www.opengis.net/def/crs/OGC/1.3/CRS84h'.
+  ///
+  /// References WGS 84 geographic coordinates (longitude, latitude) ordered as
+  /// specified by `AxisOrder.xy`, with ellipsoidal height (elevation) as a
+  /// third coordinate.
+  static const CoordRefSys CRS84h =
+      CoordRefSys.id('http://www.opengis.net/def/crs/OGC/1.3/CRS84h');
+
+  /// The coordinate reference system identified by
+  /// 'http://www.opengis.net/def/crs/EPSG/0/4326'.
+  ///
+  /// References WGS 84 geographic coordinates (latitude, longitude) ordered as
+  /// specified by `AxisOrder.yx`.
+  static const CoordRefSys EPSG_4326 =
+      CoordRefSys.id('http://www.opengis.net/def/crs/EPSG/0/4326');
+
+  /// The coordinate reference system identified by
+  /// 'http://www.opengis.net/def/crs/EPSG/0/3857'.
+  ///
+  /// References WGS 84 projected (Web Mercator) metric coordinates ordered as
+  /// specified by `AxisOrder.xy`.
+  /// 
+  /// Also known as WGS 84 / Pseudo-Mercator. Uses "spherical development of
+  /// ellipsoidal coordinates".
+  static const CoordRefSys EPSG_3857 =
+      CoordRefSys.id('http://www.opengis.net/def/crs/EPSG/0/3857');
+
+  /// The coordinate reference system identified by
+  /// 'http://www.opengis.net/def/crs/EPSG/0/3395'.
+  ///
+  /// References WGS 84 projected (World Mercator) metric coordinates ordered as
+  /// specified by `AxisOrder.xy`.
+  static const CoordRefSys EPSG_3395 =
+      CoordRefSys.id('http://www.opengis.net/def/crs/EPSG/0/3395');
+
+  @override
+  String toString() => id;
+
+  @override
+  bool operator ==(Object other) => other is CoordRefSys && id == other.id;
+
+  @override
+  int get hashCode => id.hashCode;
+}